+++ uce.portal/uce.common/src/main/java/org/texttechnologylab/uce/common/config/SpringConfig.java
@@ -3,16 +3,7 @@ package org.texttechnologylab.uce.common.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.texttechnologylab.uce.common.services.AuthenticationService;
-import org.texttechnologylab.uce.common.services.GbifService;
-import org.texttechnologylab.uce.common.services.GoetheUniversityService;
-import org.texttechnologylab.uce.common.services.JenaSparqlService;
-import org.texttechnologylab.uce.common.services.LexiconService;
-import org.texttechnologylab.uce.common.services.MapService;
-import org.texttechnologylab.uce.common.services.PostgresqlDataInterface_Impl;
-import org.texttechnologylab.uce.common.services.RAGService;
-import org.texttechnologylab.uce.common.services.S3StorageService;
-import org.texttechnologylab.uce.common.services.WikiService;
+import org.texttechnologylab.uce.common.services.*;
 
 @Configuration
 @Import({DocumentAccessConfig.class})
@@ -59,6 +50,11 @@ public class SpringConfig {
     @Bean
     public RAGService ragService() {
         return new RAGService(databaseService());
+    }
+
+    @Bean
+    public EmbeddingService embeddingService() {
+        return new EmbeddingService(databaseService());
     }
 
     @Bean
